+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "octocrab"]
-	path = octocrab
+	path = cli/octocrab
 	url = https://github.com/PurdueSoftEng/octocrab.git
