+++ Cargo.toml
@@ -1,25 +1,9 @@
 [package]
-name = "tool"
+name = "module-registry"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-octocrab = {path="./octocrab"}
-clap = { version = "4.0", feature = ["derive"], features = ["derive"] }
-anyhow = "1.0"
-log = "0.4.17"
-env_logger = "0.10.0"
-tokio = { version = "1.17.0", default-features = false, features = [
-    "macros",
-    "rt-multi-thread",
-    "time",
-] }
-serde = "1.0.152"
-chrono = "0.4.23"
-dotenv = "0.15.0"
-serde_json = "1.0.92"
-regex = "1.7.1"
-log4rs = "1.2.0"
-tempfile = "3.3.0"
+cli = { version = "0.1.0", path = "cli" }
