+++ cli/scripts/.rank.sh
