+++ cli/scripts/.test.sh
