+++ cli/scripts/install.sh
