+++ cli/scripts/rank.sh
