+++ cli/src/calc_license.rs
