+++ cli/src/calc_net_score.rs
