+++ cli/src/lib.rs
@@ -1,0 +1,3 @@
+pub fn working() -> bool {
+    true
+}
